requirements.txt
langchain
langchain_community
ipykernel
python-dotenv
pypdf
beautifulsoup4
langchain-text-splitters
faiss-cpu
langchain_chroma
langchain-openai
langchain_huggingface
streamlit
langchain_groq
langchain_core
uvicorn
langserve
fastapi
transformers 
unstructured
arxiv
wikipedia
validators==0.28.1
youtube_transcript_api
pytube